Use VERSION file for version detection
Instead of calling the go binary with the version subcommand and extract the version information from there, the VERSION file, distributed with each Go installation, is used to extract the version of a Go installation.

This is preferable, since some antivirus programs might get triggered if our process calls other programs right after the start. It should also come with less overhead and seems more robust.